+++ tailwind.config.ts
@@ -69,6 +69,15 @@ export default {
       fontFamily: {
         sans: ["var(--font-inter)"],
       },
+      keyframes: {
+        "modal-in": {
+          "0%": { opacity: "0", transform: "translateY(10px)" },
+          "100%": { opacity: "1", transform: "translateY(0)" },
+        },
+      },
+      animation: {
+        "modal-in": "modal-in 0.3s ease-out",
+      },
     },
   },
   plugins: [require("tailwindcss-animate")],
