--- types/port.ts
@@ -1,4 +1,0 @@
-export interface PortOption {
-  value: string;
-  label: string;
-}
